MAINT: Moving calibration decorator to the lr58 experiment file
